requirements.txt
alembic
annotated-types
anyio
asgiref
certifi
cffi
charset-normalizer
click
colorama
colorlog
contourpy
cryptography
cycler
defusedxml
deprecation
distlib
Django
filelock
fonttools
fsspec
gotrue
greenlet
gunicorn
h11
httpcore
httpx
idna
intel-openmp
Jinja2
joblib
kiwisolver
Mako
MarkupSafe
matplotlib
mkdocs-material-extensions
mkl
mpmath
napas-qr-python
networkx
nltk
numpy
oauthlib
opencv-python-headless
packaging
pandas
pillow
platformdirs
postgrest
psycopg2
psycopg2-binary
pycparser
pydantic
pydantic_core
PyJWT
pyparsing
pypng
python-dateutil
python-dotenv
python3-openid
pytz
PyYAML
qrcode
realtime
regex
requests
requests-oauthlib
segno
six
sniffio
social-auth-app-django
social-auth-core
SQLAlchemy
sqlparse
storage3
StrEnum
supabase
supafunc
sympy
tbb
torch
torchvision
tqdm
types-requests
typing_extensions
tzdata
Unidecode
urllib3
virtualenv
websockets
whitenoise
